Using long-gauge for container-startuptime

+++ samza-core/src/test/scala/org/apache/samza/container/TestSamzaContainer.scala
@@ -80,7 +80,7 @@ class TestSamzaContainer extends AssertionsForJUnit with MockitoSugar {
   def setup(): Unit = {
     MockitoAnnotations.initMocks(this)
     setupSamzaContainer(Some(this.applicationContainerContext))
-    when(this.metrics.containerStartupTime).thenReturn(mock[Gauge[Float]])
+    when(this.metrics.containerStartupTime).thenReturn(mock[Gauge[Long]])
   }
 
   @Test
